requirements.txt
bcrypt==4.1.3
findspark==2.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
pandas==2.2.2
pyspark==3.2.1
python-dotenv==1.0.1
setuptools==70.0.0
Werkzeug==3.0.3
tensorflow == 2.11.0
numpy == 1.24.3
scipy == 1.9.0
scikit-learn == 1.2.0
